Optimize transfers point using priority
The OptimizeTransferService is run AFTER the Raptor search and find all
possible combination of transfers for all paths. This allow us to
compare the possible transfers for a path and pick the best option. We
can filter the paths based on a optimal-transfer-cost, remove
"back-travel-options" and implement a transfer priority(GTFS transfers
and NeTEx Interchanges).