refactor: Use stop name and index in logging/debugging
Raptor uses the stop index to reference stops and avoid using references or store any other
data about a stop. This is inconvenient when logging and debugging. So to improve this, we
inject a service for resolving the stop name when a stop is logged. This is also used for
debugging stops and in toStrings().